Add support for splitting on specific trinkets, add option to only split on game complete if all trinkets are collected
Functionality has been added to allow splitting on specific trinkets and also allowing splitting on Victoria's trinket which I believe wasn't possible with the previous method of splitting on trinkets. 

Additionally, functionality has been added to prevent splitting on game completion unless all 20 trinkets are collected. This is useful in particular for 100% as it removes the need for an extra split to deal with there being two Game Complete sequences in the current RTA route.